Contact included and other more funs

+++ admin/includes/view_all_users.php
@@ -134,12 +134,20 @@ if(isset($_GET['change_to_sub'])) {
 
 if(isset($_GET['delete'])) {
 
+    if(isset($_SESSION['user_role'])) {
 
-    $the_user_id = $_GET['delete'];
+        if ($_SESSION['user_role'] == 'admin') {
 
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_user_query = mysqli_query($con, $query);
-    header("Location: users.php");
+
+
+        $the_user_id = mysqli_real_escape_string($con, $_GET['delete']);
+
+        $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+        $delete_user_query = mysqli_query($con, $query);
+        header("Location: users.php");
+       }
+
+    }
 }
 
 ?>
